+++ changelog.md
@@ -10,6 +10,7 @@
 
 ### Bugfixes
 * #694 Product variant attribute in product page should not be preselected implicitly
+* Fixed: If currencies are limited to one for a multi-store, this currency should dominate the setting for the primary store currency
 
 
 ## SmartStore.NET 2.2.1