+++ changelog.md
@@ -1,12 +1,16 @@
 ﻿# Release Notes
 
+## SmartStore.NET 2.5
+
+### New Features
+* #141 Payment and shipping methods by customer roles
+* #67 Restrict payment methods to certain countries
+* #94 Restrict payment methods to certain shipping methods
+
 ## SmartStore.NET 2.2.2
 
 ### New Features
 * #210 Implement multi-store support for import/export
-* #141 Payment methods by customer roles
-* #67 Restrict payment methods to certain countries
-* #94 Restrict payment methods to certain shipping methods
 
 ### Improvements
 * License checker now supports IDN mapping for domain names